+++ internal/models/user.go
@@ -2,12 +2,10 @@ package models
 
 import "time"
 
-// User maps directly to the users table.
-// Shared by auth, post, and comment packages.
 type User struct {
-	ID        int
+	ID        int64
 	Email     string
 	Username  string
-	Password  string // bcrypt hash — never the raw password
+	Password  string
 	CreatedAt time.Time
 }
